Inventory changed from vector<Item> variable to a self class, created all working functions for Inventory. Picking up items working. Solved some bugs and Items combination.

+++ Zork/Zork/Item.cpp
@@ -16,7 +16,7 @@ void Item::setCompatibleItem(Item i) {
 }
 
 void Item::setCompatibleItems(vector<Item> v) {
-	for (int i = 0; i < v.size(); i++) {
+	for (size_t i = 0; i < v.size(); i++) {
 		this->compatibleItems.push_back(v[i]);
 	}
 }
@@ -27,7 +27,7 @@ vector<Item> Item::getCompatibleItems() {
 }
 
 bool Item::isCompatible(Item i) {
-	for (int j = 0; j < this->compatibleItems.size(); j++) {
+	for (size_t j = 0; j < this->compatibleItems.size(); j++) {
 		if (this->compatibleItems[j].name == i.name) {
 			return true;
 		}
